+++ src/Core/Backend.Core.Marten/Context/MartenEventSourcedDataContext.cs
@@ -87,6 +87,11 @@ namespace GoldenEye.Backend.Core.Marten.Context
             return eventStore.Projections.Query<TEntity>();
         }
 
+        public IQueryable<TEntity> CustomQuery<TEntity>(string query) where TEntity : class
+        {
+            return eventStore.Projections.CustomQuery<TEntity>(query).AsQueryable();
+        }
+
         public TEntity Remove<TEntity>(TEntity entity, int? version = null) where TEntity : class
         {
             return Update(entity);