+++ src/Core/Backend.Core.WebApi/Changelog.md
@@ -1,4 +1,10 @@
-﻿# v5.0.1 (12.05.2018) [Pull Request](https://github.com/oskardudycz/GoldenEye/pull/61)
+﻿# v5.2.0 (21.05.2018) [Pull Request](https://github.com/oskardudycz/GoldenEye/pull/59)
+
+## Changes
+
+* Updated reference to Backend.Core **[MINOR]**
+
+# v5.0.1 (12.05.2018) [Pull Request](https://github.com/oskardudycz/GoldenEye/pull/61)
 
 ## Changes